some folders refactoring / fix README.md
